requirements.txt
coverage==4.2
duckling==1.0.0
JPype1==0.6.1
py==1.4.31
pytest==3.0.3
pytest-cov==2.3.1
python-dateutil==2.5.3
six==1.10.0
